packages/viewer/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "itkviewer"
description = "Python interface itk-viewer."
authors = [{name = "Matt McCormick", email = "matt.mccormick@kitware.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'Intended Audience :: Science/Research',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
keywords = [
    "itk",
    "imaging",
    "visualization",
]
dynamic = ["version"]

requires-python = ">=3.8"
dependencies = [
    "numpy",
    "ngff-zarr",
    "pydantic",
    "python-statemachine",
]

[project.urls]
Home = "https://github.com/InsightSoftwareConsorrtium/itk-viewer"
Source = "https://github.com/InsightSoftwareConsortium/itk-viewer"
Issues = "https://github.com/InsightSoftwareConsortium/itk-viewer/issues"

[tool.hatch.envs.default]
dependencies = [
    "pytest >=2.7.3",
    "pytest-pyodide",
    "itkwasm-image-io",
]

[tool.hatch.envs.default.scripts]
test = [
    "hatch build -t wheel",
    "pytest --dist-dir=./dist --rt=chrome",
]
download-pyodide = [
  "curl -L https://github.com/pyodide/pyodide/releases/download/0.25.0/pyodide-0.25.0.tar.bz2 -o pyodide.tar.bz2",
  "tar xjf pyodide.tar.bz2",
  "rm -rf dist pyodide.tar.bz2",
  "mv pyodide dist",
]

[tool.hatch.version]
path = "itkviewer/__init__.py"
